lint: consolidate ajv-related code, visitor trait interfaces
