--- src/visitor/VisitorResult.ts
@@ -1,4 +1,0 @@
-export interface VisitorResult {
-  changes: ReadonlyArray<any>;
-  errors: ReadonlyArray<any>;
-}
